restructure schedules
