+++ src/AlgebraicRewriting.jl
@@ -3,13 +3,11 @@ using Reexport
 
 include("categorical_algebra/CategoricalAlgebra.jl")
 include("rewrite/Rewrite.jl")
-include("Schedules.jl")
-include("Visuals.jl")
+include("schedules/Schedules.jl")
 
 @reexport using .CategoricalAlgebra
 @reexport using .Rewrite
 @reexport using .Schedules
-@reexport using .Visuals
